requirements.txt
langchain==0.3.7
langchain-experimental==0.3.2
python-dotenv==1.0.1
streamlit==1.39.0
tiktoken==0.7.0
faiss-cpu==1.8.0
protobuf==5.27.3
mysql-connector-python==9.1.0
pymysql==1.1.1
sentence-transformers==3.2.1
chromadb==0.5.13
openai==1.52.2

